adapted for changes to introduce 'register file' layer
